requirements.txt
# Core Data Science & ML
tensorflow
numpy
pandas
scikit-learn
opencv-python
Pillow
matplotlib
seaborn
keras
dvc

# Deep Learning & Transfer Learning
tensorflow-hub

# API & Serving
fastapi
uvicorn
python-multipart
pydantic

# Experiment Tracking
mlflow
mlflow[extras]

# Dataset Versioning
dvc
dvc[gs]

# Monitoring & Metrics
prometheus-client

# Testing
pytest
pytest-cov

# Data Augmentation
albumentations

# Utilities
python-dotenv
pytz
kaggle>=1.5.16
tqdm